output issue.
Code is correct, but rendering issue by the library. This is the fault of the 'vpython' library. So till then think before writing the code. Expected outputs are 3d objects but given outputs are black screen. Issue is posted in their git repo, link: https://github.com/vpython/glowscript/issues/182